Made Gallery Page (BASIC) and changes
Moved the background website design out of folder and edited the relevant codes so that it's easier to find.

Created the polaroid gallery page with the images. (The photos are repeated - will need to find other photos to replace later on).

Need to do:
Make the photos clickable that will lead to a profile page.
Figure out how to add a heart button icon